Add files via upload
Exercise 4 -The owner wishes to have a report displaying a summary of Orders with their corresponding list of distinct invoices and their total amount to be able to reconcile his orders and payments. The report shall contain the details as per table below ordered by latest Order Date on top. Implement a Stored Procedure or Function to return the required information.